pyproject.toml
[project]
name = "ibmcloud-agents"
version = "0.1.0"
description = "IBMCloud Agents (A2A & MCP-compliant)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.9.0",
    "google-adk>=0.2.0",
    "a2a-server @ file:///Users/chrism1/Code/ai/agentic/a2a/a2a-server",
    "chuk-llm>=0.9.9",
    "litellm>=1.67.0",
    "python-dotenv>=1.1.0",
]
# Optional dependency groups (development)
[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.12.0",
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    "boto3>=1.34.0",
]
test = [
    "pytest>=8.0.0", 
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.12.0",
]

# Convenience meta-extras
#all = ["ibmcloud_agents"]
#dev-all = ["ibmcloud_agents[dev]"]

# --------------------------------------------------------------------
#  Authors and URLs
# --------------------------------------------------------------------
[[project.authors]]
name = "Christopher C Mitchell"
email = "redacted@ibm.com"

[project.urls]
Homepage = "https://github.com/ccmitchellusa/ibmcloud-agents"
Documentation = "https://github.com/ccmitchellusa/ibmcloud-agents"
Repository = "https://github.com/ccmitchellusa/ibmcloud-agents"
"Bug Tracker" = "https://github.com/ccmitchellusa/ibmcloud-agents/issues"
Changelog = "https://github.com/ccmitchellusa/ibmcloud-agents/blob/main/CHANGELOG.md"

# --------------------------------------------------------------------
#  🔧 setuptools-specific configuration
# --------------------------------------------------------------------
[tool.setuptools]
include-package-data = true          # ensure wheels include the data files

# Automatic discovery: keep every package that starts with "ibmcloud-agents"
[tool.setuptools.packages.find]
where = ["src"]
include = ["ibmcloud*", "supervisor_agent", "common"]
exclude = ["tests*"]

##  Runtime data files  ------------------------------------------------
#  - py.typed  -> advertises inline type hints (PEP 561)
#  - static/*  -> CSS/JS for the admin UI
#  - templates -> Jinja2 templates shipped at runtime
#[tool.setuptools.package-data]
#ibmcloud-base-agent = [
#  "py.typed",
#]
# --------------------------------------------------------------------
#  🛠  Tool configurations (black, mypy, etc.)
# --------------------------------------------------------------------
[tool.pytype]
# Directory-specific options:
inputs = ["src/ibmcloud*", "src/supervisor_agent", "src/common"]
python_version = "3.12"          # match default runtime

[tool.check-manifest]
ignore = [
    "docs/**",
    "tests/**",
    ".github/**",
    "Makefile",
]

[tool.black]
line-length = 200
target-version = ["py310", "py311", "py312"]
include = "\\.pyi?$"

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
strict = true
show_error_codes = true
warn_unreachable = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov=ibmcloud-agents"
testpaths = [ "tests",]
asyncio_mode = "auto"

# ── fawltydeps ─────────────────────────────────────────────────────
[tool.fawltydeps]
# only parse main pyproject.toml
deps = ["pyproject.toml"]

# ignore 'dev' extras so they won't show up in fawltydeps
ignore_unused = [
  "autoflake",
  "argparse-manpage",
  "bandit",
  "black",
  "bump2version",
  "check-manifest",
  "code2flow",
  "cookiecutter",
  "coverage",
  "coverage-badge",
  "darglint",
  "flake8",
  "fawltydeps",
  "gprof2dot",
  "gunicorn",
  "importchecker",
  "isort",
  "ty",
  "mypy",
  "pexpect",
  "pip-licenses",
  "pip_audit",
  "pre-commit",
  "pydocstyle",
  "pylint",
  "pylint-pydantic",
  "pyre-check",
  "pyright",
  "pyroma",
  "pyspelling",
  "pytest",
  "pytest-asyncio",
  "pytest-cov",
  "pytest-examples",
  "pytest-md-report",
  "pytest-rerunfailures",
  "pytest-xdist",
  "pytype",
  "radon",
  "ruff",
  "settings-doc",
  "snakeviz",
  "types-tabulate",
  "twine",
  "uvicorn"
]

